# [LinearAlgebra] Speedup accumulation on BTDMatrix