Validated with the following benchmark. The benchmark `BM_BTDMatrix_addBlock` uses the new implemented accumulation function and must be compared to `BM_BTDMatrix_add`.

```
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
BM_BTDMatrix_add<6, SReal>/64              6212 ns         3069 ns       224000
BM_BTDMatrix_add<6, SReal>/512            49857 ns        32087 ns        22400
BM_BTDMatrix_add<6, SReal>/1024          103536 ns        54408 ns        11200
BM_BTDMatrix_addBlock<6, SReal>/64          602 ns          391 ns      1600000
BM_BTDMatrix_addBlock<6, SReal>/512        5607 ns         3432 ns       186667
BM_BTDMatrix_addBlock<6, SReal>/1024      12176 ns         6417 ns       112000

```

https://github.com/alxbilger/SofaBenchmark/pull/36


______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).