@fredroy 2023-10-23
> Wow, @fredroy and @epernod might enjoy it!

I tried on the 3instrument_collis (BeamAdapter) and there is no noticeable speed-up (or speed-down for that matter) ; but maybe this is not a good use case