@alxbilger 2023-10-23
> I tried on the 3instrument_collis (BeamAdapter) and there is no noticeable speed-up (or speed-down for that matter) ; but maybe this is not a good use case

This is what I expected. This kind of speed up is significant with a high number of DoFs, and it's not the case when using beams. I expect that you save a few microseconds by time step, which is really not significant.