@fredroy 2023-10-24
> > I tried on the 3instrument_collis (BeamAdapter) and there is no noticeable speed-up (or speed-down for that matter) ; but maybe this is not a good use case
> 
> I am not sure that this new function is called. I think #4253 must be merged first.


Indeed ! 🚀

master; 
`[BatchGUI] 5000 iterations done in 66.0261 s ( 75.7276 FPS)`

master+ this branch + #4253 : 
`[BatchGUI] 5000 iterations done in 58.863 s ( 84.943 FPS)`