@alxbilger 2023-10-24
@fredroy thanks for the benchmark. I am pleased with the results.